# server/docs/swagger.yaml
basePath: /
definitions:
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  item.CreateRequest:
    properties:
      description:
        type: string
      epic_id:
        type: string
      estimate_minutes:
        type: integer
      priority:
        type: integer
      project_id:
        type: string
      sprint_id:
        type: string
      title:
        type: string
    type: object
  item.Item:
    properties:
      board_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      display_id:
        type: string
      epic_color:
        type: string
      epic_id:
        type: string
      epic_name:
        description: Enriched (computed from JOINs, not stored)
        type: string
      estimate_minutes:
        type: integer
      id:
        type: string
      position:
        type: integer
      priority:
        type: integer
      project_id:
        type: string
      sprint_id:
        type: string
      status_id:
        type: string
      time_in_status_minutes:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  item.ListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/item.Item'
        type: array
    type: object
  item.MoveRequest:
    properties:
      status_id:
        type: string
    type: object
  item.ReorderRequest:
    properties:
      ids:
        items:
          type: string
        type: array
      status_id:
        type: string
    type: object
  item.StatusTransition:
    properties:
      from_status_id:
        type: string
      id:
        type: string
      item_id:
        type: string
      to_status_id:
        type: string
      transitioned_at:
        type: string
    type: object
  item.TransitionsResponse:
    properties:
      transitions:
        items:
          $ref: '#/definitions/item.StatusTransition'
        type: array
    type: object
  item.UpdateRequest:
    properties:
      clear_epic:
        type: boolean
      clear_sprint:
        type: boolean
      description:
        type: string
      epic_id:
        type: string
      estimate_minutes:
        type: integer
      priority:
        type: integer
      sprint_id:
        type: string
      title:
        type: string
    type: object
...
host: api.cumin.ashinsabu.com
info:
  contact: {}
  description: OSS sprint board API
  title: Cumin API
  version: "1.0"
paths:
...
  /api/items:
    get:
      parameters:
      - description: Filter by sprint ID
        in: query
        name: sprint_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.ListResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: List items
      tags:
      - items
    post:
      consumes:
      - application/json
      parameters:
      - description: Item to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/item.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.Item'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Create item
      tags:
      - items
  /api/items/{id}:
    delete:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Delete item
      tags:
      - items
    get:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.Item'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get item
      tags:
      - items
    patch:
      consumes:
      - application/json
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/item.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.Item'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Update item
      tags:
      - items
  /api/items/{id}/move:
    post:
      consumes:
      - application/json
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Target status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/item.MoveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.Item'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Move item
      tags:
      - items
  /api/items/{id}/restore:
    post:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Restore item
      tags:
      - items
  /api/items/{id}/transitions:
    get:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.TransitionsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: List item transitions
      tags:
      - items
  /api/items/backlog:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.ListResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: List backlog items
      tags:
      - items
  /api/items/reorder:
    put:
      consumes:
      - application/json
      parameters:
      - description: Status and ordered item IDs
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/item.ReorderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.ListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Reorder items
      tags:
      - items
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: token
    type: apiKey
swagger: "2.0"